\name{ILI}
\alias{ILI}
\docType{data}

\title{Florida influenza-like illness data} 

\description{
Daily influenza-like illness (ILI) incidence rates at 67 Florida counties during years 2012-2014. The ILI incidence rates were collected by the Electronic Surveillance System for the Early Notification of Community-based Epidemics (ESSENCE) that was developed by the Florida Department of Health. Researchers can have an access to the ESSENCE database after a proper online registration. Moreover, some weather conditions during years 2012-2014 can be obtained from  the official website of the National Oceanic and Atmospheric Administration of the United States. The ILI dataset used here contains 8 variables, 
including \code{County}, \code{Date}, \code{Lat} (latitude), \code{Long} (longitude), \code{Time}, \code{Rate} (ILI incidence rate), \code{Temp} (temperature) 
and \code{RH} (relative humidity), from the two databases mentioned above. 
}
\usage{data(ILI)}
\format{A dataframe containing \eqn{N=73,432} observations of 8 variables.}


\keyword{dataset}
