% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25_dat.R
\docType{data}
\name{pm25_dat}
\alias{pm25_dat}
\title{PM2.5 concentration data}
\format{
A dataframe containing \eqn{N=200,385} observations of 6 variables.
}
\usage{
data(pm25_dat)
}
\description{
Daily PM2.5 concentration levels at 183 major cities in China during years 
2014-2016. This dataset was collected by the China National Environmental 
Monitoring Centre (CNEMC). It can be downloaded directly from the CNEMC 
offical web page. The PM2.5 dataset used here contains 6 variables, including 
\code{Year}, \code{Time}, \code{Long} (longitude), \code{Lat} (latitude), 
\code{City}, and \code{PM2.5}.
}
\author{
Kai Yang \email{kayang@mcw.edu} and Peihua Qiu
}
\keyword{pm25_dat}
